Realtor.com API pull May 2, 2021
All members of the team have a copy of the API key local.  This is the working API pull from www.Realtor.com. The returned results are in json. We parsed dict to data for results. The results "key, value" pair are converted to dataframe and also .csv file.  NO COLUMNS or DATA has been removed from the results dict.